+++ tket/src/Transformations/Replacement.cpp
@@ -16,9 +16,11 @@
 
 #include "Circuit/CircPool.hpp"
 #include "Circuit/CircUtils.hpp"
+#include "Circuit/Circuit.hpp"
 #include "Decomposition.hpp"
 #include "Gate/GatePtr.hpp"
 #include "Gate/GateUnitaryMatrix.hpp"
+#include "OpType/OpType.hpp"
 #include "OpType/OpTypeInfo.hpp"
 #include "Transform.hpp"
 
@@ -26,8 +28,8 @@ namespace tket {
 
 using namespace Transforms;
 
-static Circuit multi_controlled_to_2q(
-    const Op_ptr op, const OpType& two_q_type) {
+Circuit multi_controlled_to_2q(
+    const Op_ptr op, const std::optional<OpType>& two_q_type) {
   unsigned n_qubits = op->n_qubits();
   OpType optype = op->get_type();
   Circuit c(n_qubits);
@@ -65,9 +67,11 @@ static Circuit multi_controlled_to_2q(
       throw BadOpType("The operation is not multi-controlled", optype);
   }
 
-  if (two_q_type == OpType::CX) {
+  if (two_q_type == std::nullopt) {
+    return c;
+  } else if (two_q_type.value() == OpType::CX) {
     decompose_multi_qubits_CX().apply(c);
-  } else if (two_q_type == OpType::TK2) {
+  } else if (two_q_type.value() == OpType::TK2) {
     decompose_multi_qubits_TK2().apply(c);
   } else {
     throw BadOpType("The target 2-q gate can only be CX or TK2", optype);
@@ -114,6 +118,11 @@ Circuit CX_ZX_circ_from_op(const Op_ptr op) {
     throw BadOpType(
         "Can only build replacement circuits for basic gates", desc.type());
   switch (desc.type()) {
+    case OpType::Phase: {
+      Circuit replacement(0);
+      replacement.add_phase(op->get_params()[0]);
+      return replacement;
+    }
     case OpType::Z: {
       Circuit replacement(1);
       replacement.add_op<unsigned>(OpType::Rz, 1., {0});